softwarebot0.5-github/requirements.txt
-f /usr/share/pip-wheels
discord.py==1.7.3

